test: extract skill registry white-box suite
