+++ crates/agent-engine/src/skills/registry/tests.rs
@@ -1,0 +1,2 @@
+include!("tests/discovery_and_metadata.inc.rs");
+include!("tests/execution_resolution.inc.rs");
